prevent mitm proxy auth to be escaped prior to base64 encoding it
https://cs.opensource.google/go/go/+/refs/tags/go1.20:src/net/url/url.go;l=420-431
shows that the auth info (both username and password) gets escaped,
which is undesirable as their escape is also escaping legit chars
such as '!'.

This commits disables the escaping completly. We can also define our own escaping
if you think this is useful, but I doubt that for our use case we really
need to worry about this and can expect the user to pass
in valid credentials. Escaping would either way not help here
as you'll just replace one error (invalid chars) into
another error (not authorized)